% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{Add historical data from FAOSTAT to the GFPM simulations}
\alias{Add historical data from FAOSTAT to the GFPM simulations}
\title{This script is only working on the development machine
Update needed see}
\description{
Calculate consumption and net trade for the FAOSTAT data
 TODO: add historical price in 2010 USD,  
return A table coutaining both historical and simulated data 
prepare historical data to display along GFPM simulations

Calculate consumption and net trade
Keep this function so that
the change of NA values to 0 only affects this calculation
In other words, NA values for production and trade volumes
are kept in the FAO table
}

