% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand.R
\name{adddemandelast}
\alias{adddemandelast}
\title{Add new demand elasticities to the GFPM}
\usage{
adddemandelast(demandelast, gfpmdemand)
}
\arguments{
\item{demandelast}{a table of demand elasticities 
columns should be called "Product", "gdpelastpc" and "priceelastpc".}

\item{gfpmdemand}{data frame imported from a demand sheet from the GFPM's World.xls file}
}
\value{
a dataframe containing the GFPM demand sheet with 
new elasticities. This can be written to a csv and then copied to excel
preferable on a windows machine with Excel to keep the structure
of World.xlsx
}
\description{
Add new demand elasticities to the GFPM
}

