% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{addProductAndCountry}
\alias{addProductAndCountry}
\title{Add products, country and regions names}
\usage{
addProductAndCountry(df)
}
\arguments{
\item{df}{data frame}
}
\value{
a data frame with products, country and regions names added
}
\description{
Add products, country and regions names
}

