% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{calculateConsumptionNetTrade}
\alias{calculateConsumptionNetTrade}
\title{This script is only working on the development machine
Update needed see 
prepare historical data to display along GFPM simulations}
\description{
Calculate consumption and net trade
Keep this function so that
the change of NA values to 0 only affects this calculation
In other words, NA values for production and trade volumes
are kept in the FAO table
}

