% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean}
\usage{
clean(fileName, scenario_name, path = "rawdata/")
}
\arguments{
\item{fileName}{name of the raw data file}

\item{name}{given to the scenario}
}
\value{
A list of tables for that scenario
}
\description{
loads a PELPS data frame from the file "scenario_name" for a scenario
}

