% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{copy_pelps_folder}
\alias{copy_pelps_folder}
\title{Copy GFPM output data}
\usage{
copy_pelps_folder(scenario_name, pelps_folder = "C:/PELPS/pelps/",
  compression = "none")
}
\arguments{
\item{scenario_name}{character string giving a folder or archive name under which the scenario will be saved}

\item{pelps_folder}{folder unde which the PELPS data is saved defaults "C:/PELPS/pelps/"}

\item{compression}{character string giving the type of compression to be used (default none).}
}
\description{
This function is not needed if you copy the folder by hand to rawdata
}

