% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_countrycodes.R
\docType{data}
\name{countrycodes}
\alias{countrycodes}
\title{Country names and codes}
\format{a data frame with country codes and names}
\usage{
countrycodes
}
\description{
Country names and codes
}
\examples{
countrycodes
}

