% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/func.R
\name{load_and_clean_gfpm_data}
\alias{load_and_clean_gfpm_data}
\title{Load and clean gfpm data}
\usage{
load_and_clean_gfpm_data(scenario_name, pelps_folder = "C:/PELPS/pelps/",
  compression = "none")
}
\arguments{
\item{scenario_name}{string, name of a scenario}

\item{pelps_folder}{path to the pelps folder on windows}

\item{compression,}{see the functions copy_pelps_folder and savePELPSToRdata}
}
\description{
Load and clean gfpm data
}

