% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plotprodbycounty}
\alias{plotprodbycounty}
\title{Plot by GFPM countries}
\usage{
plotprodbycounty(scenarios, product, scenario, country)
}
\arguments{
\item{scenarios}{a list of scenarios, output of \code{\link{savePELPSToRdata}()}}

\item{product}{a character vector of product names to select}

\item{scenario}{a character vector of scenario names to select, only one scenario}

\item{country}{a character vector of countries to select}
}
\description{
Plot by GFPM countries
}

