% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/func.R
\name{plotProdByReg}
\alias{plotProdByReg}
\title{Plot Summary by GFPM regions}
\usage{
plotProdByReg(scenarios, product = "", scenario = "")
}
\arguments{
\item{scenarios}{a list of scenarios, output of \code{\link{savePELPSToRdata}()}}

\item{product}{a vector of product names to select}

\item{sceario}{a vector of scenario names to select}
}
\description{
Plot Summary by GFPM regions
}

