% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_productcodes.R
\docType{data}
\name{productcodes}
\alias{productcodes}
\title{Product names and codes}
\format{a data frame with country codes and names}
\usage{
productcodes
}
\description{
Product names and codes
}
\examples{
productcodes
}

