% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{readPELPSTable}
\alias{readPELPSTable}
\title{Read .DAT files from folder or compressed zip or bzip2 archives}
\usage{
readPELPSTable(scenario_name, fileName, compressed = "none")
}
\arguments{
\item{scenario_name}{name of a scenario}

\item{fileName}{file name}

\item{compressed}{"none" for no compression, "zip" or "bzip2"}
}
\description{
After running a scenario copy the "C:\\PELPS\\pelps" directory containing .DAT files
and rename the directory to a unique name identifying your scenario
This function will convert all interesting .DAT files to R data frames
It's also possible to compress the folder as .zip format.
}

