% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{reshapeLong}
\alias{reshapeLong}
\title{Reshape PELPS tables and extract code in a truly Long format #}
\usage{
reshapeLong(df.PELPS, elementName = "")
}
\arguments{
\item{df.PELPS}{}
}
\description{
Reshape PELPS tables and extract code in a truly Long format #
}

