% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{savePELPSToRdata}
\alias{savePELPSToRdata}
\title{Load a scenario and save it in a Rdata file}
\usage{
savePELPSToRdata(scenario_name, compressed = "none")
}
\arguments{
\item{scenario_name}{name of a scenario}

\item{compressed,}{see \code{\link{readPELPSTable}()}}
}
\value{
A list of scenarios, saved in a RDATA file.
}
\description{
Wrapper function that calls \code{\link{readPELPSTable}()}
to load the interesting .DAT files from the archive.
}

