% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{splittrade}
\alias{splittrade}
\title{Split trade in 2 dataframes for import and export #}
\usage{
splittrade(P)
}
\arguments{
\item{P}{data frame of pelps data}
}
\value{
a list of data frames
}
\description{
Split trade in 2 dataframes for import and export #
}

