% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{bindScenarios}
\alias{bindScenarios}
\title{Bind scenarios together and save an alternative output}
\usage{
bindScenarios(scenario1, scenario2)
}
\arguments{
\item{scenario1}{scenario object created
by the \code{\link{clean}()} function or
by the \code{\link{load_and_clean_gfpm_data}()} functions}

\item{scenario2}{another scenario object}
}
\value{
List of data frames
}
\description{
It is fine to analysing scenarios one by one.
But when comparing scenarios,
it is sometimes preferable to have all scenarios
in one data frame.
Since scenario data is storred in a list of 4 data frames,
this function does the job of binding each data frame in turn.
}

