% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{load_and_clean_gfpm_data}
\alias{load_and_clean_gfpm_data}
\title{Load and clean the last GFPM scenario from the PELPS folder}
\usage{
load_and_clean_gfpm_data(scenario_name, pelps_folder = "C:/PELPS/pelps/",
  compression = "none")
}
\arguments{
\item{scenario_name}{string, name of a scenario}

\item{pelps_folder}{path to the pelps folder on windows}

\item{compression,}{see the functions copy_pelps_folder and savePELPSToRdata}
}
\description{
Calls the functions \code{\link{copy_pelps_folder}()}
to copy the raw PELPS data in the ./rawdata folder.
Calls the function \code{\link{savePELPSToRdata}()}
to clean the data and save it to an .RDATA file in the ./rawdata folder.
Calls the function \code{\link{clean}()} to clean the data.
Then save the cleaned data as an .rds file in the enddata folder.
The RDS format is explained in \code{\link{saveRDS}()}.
If the ./rawdata and ./enddata folder do not exists, they will
be created under the working directory.
}

